+++ server.go
@@ -1,35 +1,13 @@
 package main
 
 import (
-	"bytes"
 	"fmt"
-	"io/ioutil"
-	"runtime/debug"
 	"strings"
-	"time"
 
 	"github.com/gin-gonic/gin"
-	"github.com/mike-webster/repo-watcher/env"
 	webhookmodels "github.com/mike-webster/repo-watcher/webhookmodels"
 	"github.com/sirupsen/logrus"
 )
-
-// CodeOK is for a 200 response
-const CodeOK int = 200
-
-// CodeNoContent is for a 204 response
-const CodeNoContent int = 204
-
-// CodeInvalid is for a 400 response
-const CodeInvalid int = 400
-
-// CodeUnauth is for a 401 response
-const CodeUnauth int = 401
-
-const errInvalidSecret = "Invalid request secret"
-const errMissingEvent = "Missing event value"
-const errInvalidBody = "Invalid POST body"
-const errInvalidHeader = "Invalid request headers"
 
 type ApiServer interface {
 	Start() error
@@ -249,112 +227,4 @@ func parseEventMessage(ctx *gin.Context, eventName string, logger *logrus.Logger
 	}
 
 	return fmt.Sprint(name, " ", event.ToString()), event.Repository(), nil
-}
-
-func requestLogger() gin.HandlerFunc {
-	return func(ctx *gin.Context) {
-		defer func(ctx *gin.Context) logrus.FieldLogger {
-			if ctx.Request.URL.Path == "/" && ctx.Writer.Status() == 200 {
-				return nil
-			}
-
-			// log body if one is given]
-			strBody := ""
-			body, err := ioutil.ReadAll(ctx.Request.Body)
-			if err != nil {
-				defaultLogger(ctx).WithField("error", err).Error("cant read request body")
-			} else {
-				// write the body back into the request
-				ctx.Request.Body = ioutil.NopCloser(bytes.NewBuffer(body))
-
-				strBody = string(body)
-				strBody = strings.Replace(strBody, "\n", "", -1)
-				strBody = strings.Replace(strBody, "\t", "", -1)
-			}
-
-			logger := defaultLogger(ctx).WithFields(logrus.Fields{
-				"client_ip":    ctx.ClientIP(),
-				"event":        "http.in",
-				"method":       ctx.Request.Method,
-				"path":         ctx.GetString("originalPath"),
-				"query":        ctx.Request.URL.RawQuery,
-				"referer":      ctx.Request.Referer(),
-				"status":       ctx.Writer.Status(),
-				"user_agent":   ctx.Request.UserAgent(),
-				"git_event":    ctx.Request.Header.Get("X-GitHub-Event"),
-				"request_body": strBody,
-			})
-
-			if len(ctx.Errors) > 0 {
-				logger.Error(strings.TrimSpace(ctx.Errors.String()))
-			} else {
-				logger.Info()
-			}
-			return logger
-		}(ctx)
-	}
-}
-
-func defaultLogger(ctx *gin.Context) *logrus.Logger {
-	if ctx == nil {
-		return newLogger()
-	}
-
-	var logger *logrus.Logger
-	l, exists := ctx.Get("logger")
-	if !exists {
-		logger = newLogger()
-		ctx.Set("logger", logger)
-		return logger
-	}
-
-	logger, ok := l.(*logrus.Logger)
-	if !ok {
-		return newLogger()
-	}
-
-	return logger
-}
-
-func newLogger() *logrus.Logger {
-	logger := logrus.New()
-	logger.Formatter = &logrus.JSONFormatter{
-		TimestampFormat: time.RFC3339Nano,
-	}
-	if env.GetConfig().LogLevel == "debug" {
-		logger.Level = logrus.DebugLevel
-	} else {
-		logger.Level = logrus.InfoLevel
-	}
-	return logger
-}
-
-// consolidate stack on crahes
-func recovery() gin.HandlerFunc {
-	return func(c *gin.Context) {
-		defer func() {
-			if r := recover(); r != nil {
-				b, _ := ioutil.ReadAll(c.Request.Body)
-
-				defaultLogger(c).WithFields(logrus.Fields{
-					"event":    "ErrPanicked",
-					"error":    r,
-					"stack":    string(debug.Stack()),
-					"path":     c.Request.RequestURI,
-					"formbody": string(b),
-				}).Error("panic recovered")
-
-				c.AbortWithStatus(500)
-			}
-		}()
-		c.Next() // execute all the handlers
-	}
-}
-
-func setDependencies(deps *AppDependencies) gin.HandlerFunc {
-	return func(ctx *gin.Context) {
-		ctx.Set("deps", deps)
-		ctx.Set("logger", defaultLogger(ctx))
-		ctx.Next()
-	}
 }
